radeon: On DVI-I use Load Detection when EDID is bogus
Patchwork-id: 55259
O-Subject: None
Bugzilla: 873728
RH-Acked-by: Adam Jackson <ajax@redhat.com>

From: Egbert Eich <eich@suse.de>

The Radeon driver uses the analog/digital flag to determine if the
DAC or the TMDS encoder should be enabled on a DVI-I connector.
If the EDID is bogus this flag is no longer reliable. This fix
adds a fallback to DAC load detection to determine if anything
is connected to the DAC. If not and a (bogus) EDID is found it
assumes a digital display is connected.
This works around problems with some crappy IPMI devices using
Radeon ES1000.

Signed-off-by: Egbert Eich <eich@suse.de>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>