ext4: flush the i_completed_io_list during ext4_truncate
Patchwork-id: 55266
O-Subject: [RHEL6.4 PATCH 1/3] ext4: flush the i_completed_io_list during ext4_truncate
Bugzilla: 869906
CVE: CVE-2012-4508
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

From: Jiaying Zhang <jiayingz@google.com>

BREW 5053847
https://brewweb.devel.redhat.com/taskinfo?taskID=5053847

BZ 869906
https://bugzilla.redhat.com/show_bug.cgi?id=869906

Upstream 3889fd57ea3c58209354862523275774fca9db03

This is partial backport of upstream commit
3889fd57ea3c58209354862523275774fca9db03, we do not backport the rename
since we use this function from different places and we're missing
several renames of that function anyway.

The justification for this backport is that we're trying prevent
possible problem of AIO racing with truncate. As noted bellow this does
not actually fixes the issue but narrows (already quite narrow) window
of opportunity. It is also non-invasive and very straightforward
one-line change. AFAIK there is no reproducer for this particular
problem, so this is just precaution measure.

Ted first found the bug when running 2.6.36 kernel with dioread_nolock
mount option that xfstests #13 complained about wrong file size during fsck.
However, the bug exists in the older kernels as well although it is
somehow harder to trigger.

The problem is that ext4_end_io_work() can happen after we have truncated an
inode to a smaller size. Then when ext4_end_io_work() calls
ext4_convert_unwritten_extents(), we may reallocate some blocks that have
been truncated, so the inode size becomes inconsistent with the allocated
blocks.

The following patch flushes the i_completed_io_list during truncate to reduce
the risk that some pending end_io requests are executed later and convert
already truncated blocks to initialized.

Note that although the fix helps reduce the problem a lot there may still
be a race window between vmtruncate() and ext4_end_io_work(). The fundamental
problem is that if vmtruncate() is called without either i_mutex or i_alloc_sem
held, it can race with an ongoing write request so that the io_end request is
processed later when the corresponding blocks have been truncated.

Ted and I have discussed the problem offline and we saw a few ways to fix
the race completely:

a) We guarantee that i_mutex lock and i_alloc_sem write lock are both hold
whenever vmtruncate() is called. The i_mutex lock prevents any new write
requests from entering writeback and the i_alloc_sem prevents the race
from ext4_page_mkwrite(). Currently we hold both locks if vmtruncate()
is called from do_truncate(), which is probably the most common case.
However, there are places where we may call vmtruncate() without holding
either i_mutex or i_alloc_sem. I would like to ask for other people's
opinions on what locks are expected to be held before calling vmtruncate().
There seems a disagreement among the callers of that function.

b) We change the ext4 write path so that we change the extent tree to contain
the newly allocated blocks and update i_size both at the same time --- when
the write of the data blocks is completed.

c) We add some additional locking to synchronize vmtruncate() and
ext4_end_io_work(). This approach may have performance implications so we
need to be careful.

All of the above proposals may require more substantial changes, so
we may consider to take the following patch as a bandaid.

Signed-off-by: Jiaying Zhang <jiayingz@google.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Jarod Wilson <jarod@redhat.com>