cifs: extra sanity checking for cifs.idmap keys
Patchwork-id: 55315
O-Subject: [RHEL6.4 PATCH 14/16] [bz868433]cifs: extra sanity checking for cifs.idmap keys
Bugzilla: 868433
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Fixes bz868433.

Now that we aren't so rigid about the length of the key being passed
in, we need to be a bit more rigorous about checking the length of
the actual data against the claimed length (a'la num_subauths field).

Check for the case where userspace sends us a seemingly valid key
with a num_subauths field that goes beyond the end of the array. If
that happens, return -EIO and invalidate the key.

Also change the other places where we check for malformed keys in this
code to invalidate the key as well.

Reviewed-by: Shirish Pargaonkar <shirishpargaonkar@gmail.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit f09b9cbbc6999dac18436d8d4355b04d05554a48)

Also use tmp_key_invalidate() for now in place of key_invalidate() which
exists upstream as this hasn't been backported to RHEL 6 yet.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>