vfs: get rid of S_BIAS
Patchwork-id: 55374
O-Subject: [RHEL-6.4 PATCH] vfs - get rid of S_BIAS
Bugzilla: 872813
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=872813

Bug 768304 introduced a deadlock of ->s_umount.
This was due to an unmatched up_write() in fs/super.c:grab_super().

The commit below appears to be needed although it doesn't fix the
unmatched up_write() but to make grab_super() consistent with the
upstream code, as it was at the time of the patch, a down_write()
needed to be added within grab_super().

commit b20bd1a5e78af267dc4b6e1ffed48d5d776302c5

get rid of S_BIAS

    use atomic_inc_not_zero(&sb->s_active) instead of playing games with
    checking ->s_count > S_BIAS

    Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>

Signed-off-by: Ian Kent <ikent@redhat.com>