packet: fix network header offset in skb for raw sockets
Patchwork-id: 55416
O-Subject: [RHEL6.5 PATCH] net: fix network header offset in skb for raw sockets
Bugzilla: 845515
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=845515

Description:
Due to a flaw in dhcp, (and possibly other programs as well) with
handling of raw sockets, the network header offset in the skb is not
properly set. This may cause hardware to drop packets as seen in this
bugzilla. This got fixed upstream by calling skb_set_network_header() in
the AF_PACKET send path.
commit bfd5f4a3d605e0f6054df0b59fe0907ff7e696d3
Author: Sridhar Samudrala <sri@us.ibm.com>
Date:   Thu Feb 4 20:24:10 2010 -0800

    packet: Add GSO/csum offload support.

The virtio net parts of this commit have been skipped by request of mst.
Heading for a minimal fix to resolve this.

Upstream Status:
http://git.kernel.org/linus/bfd5f4a3d605e0f6054df0b59fe0907ff7e696d3

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=5076770

Test:
Tested by the reporter and myself.

Signed-off-by: Jarod Wilson <jarod@redhat.com>