gfs2: Test bufdata with buffer locked and gfs2_log_lock held
Patchwork-id: 55426
O-Subject: [RHEL 6.4 PATCH] GFS2: Test bufdata with buffer locked and gfs2_log_lock held
Bugzilla: 874319
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

In gfs2_trans_add_bh(), gfs2 was testing if a there was a bd attached to the
buffer without having the gfs2_log_lock held. It was then assuming it would
stay attached for the rest of the function. However, without either the log
lock being held of the buffer locked, __gfs2_ail_flush() could detach bd at any
time.  This patch moves the locking before the test.  If there isn't a bd
already attached, gfs2 can safely allocate one and attach it before locking.
There is no way that the newly allocated bd could be on the ail list,
and thus no way for __gfs2_ail_flush() to detach it.

This patch is the fix for BZ #874319

This patch has already been pushed upstream:

http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=96e5d1d3adf56f1c7eeb07258f6a1a0a7ae9c489

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>