internal: build a single tarball based on lastcommit
Patchwork-id: 4987
O-Subject: [kernel team] [RFC RHEL6 PATCH 1/5 V2] internal: build a single tarball based on lastcommit
Bugzilla:
RH-Acked-by: Neil Horman <nhorman@redhat.com>

This patch lays the ground work for the upcoming patches

RHEL-6 has upwards of 20K patches applied to 2.6.32.  Splitting those out and
parsing them for an srpm takes a long time (cleaning them up from a previous
run takes a long time too).  And then on the flip side, unpack the srpm on
brew and applying 20K patches with 'patch -p1' takes a long time.

We are talking 10-15 minutes on each side of the srpm (pack and unpack).

We moved to a single tarball and probably are not going back.  Let's
embace the work Aris and Jarod did on distgit and create a single
tarball for srpms.

This saves about 10-15 minutes on the build server (depending on the speed
of the machine).

To accomplish this, I will create a seperate build type instance in the
Makefile with the next patch.  As a result it won't affect today's workflow.

Instead of basing off of 2.6.32, this patch bases off of 'lastcommit' if
choosen to.  Only two places are really affected.

The first is the git-archive command and the second is in create-patches
to determine where to start popping off patches.

This is all setup to be an opt-in solution which the next patch will do.

The ideas and worked are based off of Aris and Jarod's distgit stuff and
the tweaks Kyle did for RHEL-7.

I tested by the .gitattribute stuff by building the tarball and comparing
the result to the same version under distgit.  The only differences were the
inclusion of various .gitignore files and the removal of kernel-*.config files
from the new tree.  (also some leftover wireless files were removed).

The .gitattribute trick (thanks Kyle) has 'redhat' and 'redhat/'.  My testing
has showed only 'redhat' works to ignore that directory but Kyle's testing and
some websites have shown 'redhat/' to work correctly.  I would like to find out
what version of git (I tested f17 and rhel6) needs 'redhat/' to verify that
.gitattributes still works correctly.

The reason for using the command line route is to handle the building from
Koji target seen in rhg-*.  The way koji builds the srpm makes things a
little tricky to pass internal variables around.  Plus it makes for an easy
way to test.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>