internal: implement brew speed up hacks
Patchwork-id: 4986
O-Subject: [kernel team] [RFC RHEL6 PATCH 2/5 V2] internal: implement brew speed up hacks
Bugzilla:
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Compiling on brew is slooow.  Even worse most developers just compile to
make sure their patches build on all arches.  IOW they throw away the build.

If we are going to throw it away let's restrict the pieces we build
to the pieces that are useful.  For example, processing -debuginfo is
very rarely going to fail for folks and sucks up 500 MB of disk processing
and space.  Compiling -doc rarely breaks and is serialized too.

Combined with the previous patch, we can chop off about 30-40 minutes per
arch for each build.  For most arches, that is half the build time (ppc64
is excluded in the 'half' stat).

These options will be used in subsequent patches in this series.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>