cifs: on CONFIG_HIGHMEM machines, limit the wsize to the kmap space
Patchwork-id: 55444
O-Subject: [RHEL6.4 PATCH 1/3] [bz832252]cifs: on CONFIG_HIGHMEM machines, limit the wsize to the kmap space
Bugzilla: 832252
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

This is a partial backport of commit
3ae629d98bd5ed77585a878566f04f310adbc591
which at the moment only affects async writes.
I will be backporting the remaining part of the patch once we have
included support for async reads into the RHEL 6 kernel.

--
cifs: on CONFIG_HIGHMEM machines, limit the rsize/wsize to the kmap
space

We currently rely on being able to kmap all of the pages in an async
read or write request. If you're on a machine that has CONFIG_HIGHMEM
set then that kmap space is limited, sometimes to as low as 512 slots.

With 512 slots, we can only support up to a 2M r/wsize, and that's
assuming that we can get our greedy little hands on all of them. There
are other users however, so it's possible we'll end up stuck with a
size that large.

Since we can't handle a rsize or wsize larger than that currently, cap
those options at the number of kmap slots we have. We could consider
capping it even lower, but we currently default to a max of 1M. Might as
well allow those luddites on 32 bit arches enough rope to hang
themselves.

A more robust fix would be to teach the send and receive routines how
to contend with an array of pages so we don't need to marshal up a kvec
array at all. That's a fairly significant overhaul though, so we'll need
this limit in place until that's ready.

Cc: <stable@vger.kernel.org>
Reported-by: Jian Li <jiali@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 3ae629d98bd5ed77585a878566f04f310adbc591)
--

Signed-off-by: Sachin Prabhu
Signed-off-by: Jarod Wilson <jarod@redhat.com>