sunrpc: Ensure that rpc_release_resources_task() can be called twice
Patchwork-id: 55589
O-Subject: [RHEL6.4 PATCH] BZ#878204: sunrpc: Ensure that rpc_release_resources_task() can be called twice
Bugzilla: 878204
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>

This patch fixes a nastily subtle rpc_cred refcount underflow. Several
customers have reported oopses that are apparently due to this problem.

When a new rpc_task is created, the code takes a reference to a rpc_cred
and sets the task->tk_cred pointer to it. When the call completes, the
FSM calls rpc_release_resources_task to free up the resources held by
the rpc_task, including the rpc_cred.

When it releases the rpc_cred reference however it does not zero out the
pointer to it and when the final put of the task is done it ends up
releasing that reference again. This patch fixes that by having the code
zero out the pointer.

The RHEL bug report is here:

    https://bugzilla.redhat.com/show_bug.cgi?id=878204

It's often not obvious, so tracking this down and verifying the fix
required an instrumented kernel that would catch these refcounting
underflows.

This is a long standing bug in RHEL6 and will likely need to go into
z-stream as well:

Upstream commit id: a271c5a0dea418931b6a903ef85adc30ad4c54be
Original patch description:

--------------------------[snip]--------------------------

BUG: atomic_dec_and_test(): -1: atomic counter underflow at:
Pid: 2827, comm: mount.nfs Not tainted 2.6.38 #1
Call Trace:
 [<ffffffffa02223a0>] ? put_rpccred+0x44/0x14e [sunrpc]
 [<ffffffffa021bbe9>] ? rpc_ping+0x4e/0x58 [sunrpc]
 [<ffffffffa021c4a5>] ? rpc_create+0x481/0x4fc [sunrpc]
 [<ffffffffa022298a>] ? rpcauth_lookup_credcache+0xab/0x22d [sunrpc]
 [<ffffffffa028be8c>] ? nfs_create_rpc_client+0xa6/0xeb [nfs]
 [<ffffffffa028c660>] ? nfs4_set_client+0xc2/0x1f9 [nfs]
 [<ffffffffa028cd3c>] ? nfs4_create_server+0xf2/0x2a6 [nfs]
 [<ffffffffa0295d07>] ? nfs4_remote_mount+0x4e/0x14a [nfs]
 [<ffffffff810dd570>] ? vfs_kern_mount+0x6e/0x133
 [<ffffffffa029605a>] ? nfs_do_root_mount+0x76/0x95 [nfs]
 [<ffffffffa029643d>] ? nfs4_try_mount+0x56/0xaf [nfs]
 [<ffffffffa0297434>] ? nfs_get_sb+0x435/0x73c [nfs]
 [<ffffffff810dd59b>] ? vfs_kern_mount+0x99/0x133
 [<ffffffff810dd693>] ? do_kern_mount+0x48/0xd8
 [<ffffffff810f5b75>] ? do_mount+0x6da/0x741
 [<ffffffff810f5c5f>] ? sys_mount+0x83/0xc0
 [<ffffffff8100293b>] ? system_call_fastpath+0x16/0x1b

Well, so, I think this is real bug of nfs codes somewhere. With some
review, the code

rpc_call_sync()
    rpc_run_task
        rpc_execute()
            __rpc_execute()
                rpc_release_task()
                    rpc_release_resources_task()
                        put_rpccred()                <= release cred
    rpc_put_task
        rpc_do_put_task()
            rpc_release_resources_task()
                put_rpccred()                        <= release cred again

seems to be release cred unintendedly.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>