Log thin provisioning threshold event
Patchwork-id: 55607
O-Subject: [RHEL6.4 PATCH] [SCSI] Log thin provisioning threshold event
Bugzilla: 847998
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

This is a cherry-pick of an upstream commit to log a warning message
when a storage device reports a certain SCSI Unit Attention code.

Request exception for inclusion in RHEL6.4, this is a very small change
that consists of adding a kernel printk() for a specific SCSI error code.
It will help alleviate administration difficulties for people using the
thin provisioning feature by informing them when they are close to the limit.

Bugzilla: 847998
Upstream Status: In upstream linux-stable
Brew: 5128151
Tested: Yes

Tested by modifying scsi_debug to generate the Unit Attention with
ASC = 0x38 and ASCQ = 0x07 to verify that the message appears in
/var/log/messages.