vfs: prefer EEXIST to EROFS when creating on an RO filesystem
Patchwork-id: 55636
O-Subject: [RHEL6.4 PATCH] vfs: prefer EEXIST to EROFS when creating on an RO filesystem
Bugzilla: 878091
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Dave Chinner <dchinner@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

This is for
Bug 878091 - fsck executed on underlying path instead of multipath device

The root cause of the problem was that commit
1d3c71fcd66a1204fb43f3f06221732c6df1c7c2 [fs] Push mnt_want_write() outside of i_mutex
ended up returning EROFS when i.e. a mkdir was attempted on
a readonly filesystem, but the dir already existed.  Previously
we returned EEXIST.  Either return should be acceptable, but
this changed prior behavior, and broke some applications.

Also, upstream explicitly preserved the EEXIST behavior,
see comment in kern_path_create(), so we should match that,
which is what this patch does, by stashing the mnt_want_write()
error return until after we've checked for the file already
existing.

This patch is NOT from upstream, because it's fixing a prior commit
which is also not upstream.

This also fixes a bug in symlinkat which would leave a ref on
the fs if it errored out at mnt_want_write(), leaving the fs
unable to umount.  I didn't send it as a separate patch because that
patch's contents would be immediately removed by this new change.

I've tested this on the box showing the original problem, and I
also wrote and submitted an xfstest to check this behavior for
mkdir, mknod, and symlinks.

Reviewers: please triple-check error exit behavior.  :)

Thanks,
-Eric