bna: put back VLAN tag stripped in promisc mode when no VLAN is configured
Patchwork-id: 55658
O-Subject: [RHEL6.4 PATCH] bna: put back VLAN tag stripped in promisc mode when no VLAN is configured
Bugzilla: 870279
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

BZ: #870279 - [bna] get the wrong ipv6 address from dhcp6 server on vlan

Description:
The bna driver supports automatic VLAN tag stripping as well as VLAN
filtering. The driver strips the VLAN tag _always_, this means also in
case where is no VLAN configured over the interface. In normal mode there
is no problem, because if no VLAN is configured the VLAN filter in HW
drops such packets.
But the problem occurs when the NIC is running in promiscuous mode. In
this case the filter is off and packets with VLAN tag are received by
the driver, but with stripped VLAN tags. And because there is no VLAN
configuration over the interface (vlan_grp is NULL) the driver uses
napi_gro_receive / netif_receive_skb instead of vlan_gro_receive /
vlan_hwaccel_receive_skb. The result is the stack sees these packets as
they don't have the VLAN packets. Instead of this the driver needs to
insert stripped VLAN tag back prior to push them into the stack.
The same bug was found in be2net driver in the past, the fix is the same.
Note that the upstream is not affected as it handles VLAN packets
differently.

Test status:
Tested by myself inhouse

Upstream commits:
N/A - upstream uses "new VLAN model"

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>