memcontrol: propagate LRU accounting state when splitting THP
Patchwork-id: 55742
O-Subject: Re: [PATCH RHEL6.4 BZ881714] memcg: Fix VM_BUG_ON condition in mem_cgroup_lru_del_list function
Bugzilla: 881714
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

Splitting a huge page does not transfer the page_cgroup's LRU
accounting state bit to the tail pages, even though they are accounted
to the LRU list of the memcg they are currently on.  When the page is
isolated or freed later on, they will be unaccounted from the root
memcg's LRU stats because of this, underflowing the root memcg and
leaking a page on the other memcg's LRU counter.

Correctly transfer the LRU accounting bit to all tail pages when
splitting a huge page.

Signed-off-by: Johannes Weiner <jweiner@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>