tg3: Do not set TSS for 5719 and 5720
Patchwork-id: 55750
O-Subject: [RHEL6.4 PATCH BZ823371] tg3: Do not set TSS for 5719 and 5720
Bugzilla: 823371
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

tg3: Transmit/Receive performance on BCM5719
https://bugzilla.redhat.com/show_bug.cgi?id=823371

Upstream: The basis for this patch is upstream, albeit in pieces of two patches.

Brew: Successfully built in Brew for all architectures (task_5160506)

Tested: I ran Connectathon successfully on a Beaker system with tg3 for
  sanity sake.

Note: Although a series of upstream patches could be used to refactor a couple
of functions, it was determined that doing that would involve too much code
change for a 6.4 snapshot. Broadcom returned with a proposal to disable TSS
which is reflected in the proposed patch. Broadcom's change from bz comment #22
is:
 	if (tp->irq_cnt > 1) {
 		tg3_flag_set(tp, ENABLE_RSS);

+		/* Disable multiple TX rings by default.  Simple round-robin
+		 * hardware scheduling of the TX rings can cause starvation of
+		 * rings with small packets when other rings have TSO or jumbo
+		 * packets.
+		 */
+#if 0
 		if (GET_ASIC_REV(tp->pci_chip_rev_id) == ASIC_REV_5719 ||
 		    GET_ASIC_REV(tp->pci_chip_rev_id) == ASIC_REV_5720) {
 			tg3_flag_set(tp, ENABLE_TSS);
 			netif_set_real_num_tx_queues(tp->dev, tp->irq_cnt - 1);
 		}
+#endif

Since the first upstream commit below removes the check for 5719 and 5720 and
the second removes the ENABLE_TSS and the setting of the real num queues, I
felt it was better to remove the code rather than #if it, as Broadcom proposed.
Since the comment is not upstream and the code is now gone, I did not include
the comment.

commit cf6d6ea660f46a030a5185cf577856bdd547c2bf
Author: Michael Chan <mchan@broadcom.com>
Date:   Fri, 28 Sep 2012 07:12:43 +0000 (+0000)

tg3: Disable multiple TX rings by default due to hardware flaw

Simple round-robin hardware TX scheduling can cause starvation of TX rings
with small packets when other TX rings have large TSO or jumbo packets.

In the simplest case, consider 2 TCP streams running in opposite
directions.  The TSO TX traffic will hash to one ring and the ACKs for the
incoming data on a different TCP connection will hash to a different TX
ring.  The hardware fetches one complete TSO packet (up to 64K data)
before servicing the other TX ring.  When it gets to the other TX ring, it
will only fetch one packet (64-byte ACK packet in this case).  After that,
it will switch back to the 1st ring filled with more TSO packets.  Because
only one ACK can go out roughly every 500 usec in this case, the incoming
data rate becomes very low.

Signed-off-by: Michael Chan <mchan@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>