oprofile: Fix marked events support on Power7+ not set
Patchwork-id: 55853
O-Subject: [PATCH RHEL6.5 BZ873425 2/3] powerpc/oprofile: Fix marked events support on Power7+ not set.
Bugzilla: 873425
RH-Acked-by: Don Zickus <dzickus@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=873425

Description:
------------
Starting with Power 7+ we need to check for marked events if the SIAR
register is valid, i.e. it contains the correct address of the instruction
at the time the performance counter overflowed.  The mmcra register on
Power 7+, contains a new bit to indicate that the contents of the SIAR
is valid. If the event is not marked, then the sample is recorded
independently of the SIAR valid bit setting.  For older processors, there
is no SIAR valid bit to check so the samples are always recorded.  This is
done by forcing the cntr_marked_events bit mask to zero.  The code will
always record the sample in this case since the bit mask says the event is
not a marked event even if it really is a marked event.

Signed-off-by: Carl Love <cel@us.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
6.3

kABI Status:
------------
No symbols were harmed.

Brew:
-----
Built on all platforms. All patches brewed together.

Upstream Status:
----------------
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=adbf115d66649b3a52e6875f93c9f131ba690cce

--------------------------------------------------------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>