handle null sb in get_super_thawed
Patchwork-id: 55859
O-Subject: [RHEL6.4 PATCH] fs: handle null sb in get_super_thawed
Bugzilla: 874521
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

This is for
Bug 874521 - cifs on btrfs: NULL pointer dereference when running fsstress on cifs with btrfs as backend fs

on btrfs, get_super(bdev) will return NULL.

Upstream, we just return NULL in get_super_thawed if that's
what we get:

                if (!s || s->s_writers.frozen == SB_UNFROZEN)
                        return s;

and we need to do the same thing in both sides of the
kabi-switch-enhanced version in RHEL6.

Eryu Guan proposed & tested this patch, and reported that it
works as expected.

It's not upstream, because this is all in KABI-workaround land.

Thanks,
-Eric

Signed-off-by: Jarod Wilson <jarod@redhat.com>