perf: Export Sandy Bridge uncore clockticks event in sysfs
Patchwork-id: 55861
O-Subject: [RHEL6.4 BZ886374 PATCH 2/2] perf/x86: Export Sandy Bridge uncore clockticks event in sysfs
Bugzilla: 886374
RH-Acked-by: Don Zickus <dzickus@redhat.com>

commit 35534b201c9f115c68962c095b5a9aad204d025f
Author: Stephane Eranian <eranian@google.com>
Date:   Wed Aug 29 15:01:22 2012 +0200

This patch exports the clockticks event and its encoding to user level.
The clockticks event was exported for Nehalem/Westmere but not for Sandy
Bridge (client). Given that it uses a special encoding, it needs to be
exported to user tools, so users can do:

Signed-off-by: Stephane Eranian <eranian@google.com>
Acked-by: Yan, Zheng <zheng.z.yan@intel.com>
Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
Link: http://lkml.kernel.org/r/20120829130122.GA32336@quad
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Signed-off-by: Jarod Wilson <jarod@redhat.com>