gfs2: Journal DLM lock has wrong label
Patchwork-id: 55935
O-Subject: [RHEL 6.4] GFS2: Journal DLM lock has wrong label (bz #884822)
Bugzilla: 884822
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This is a one-liner patch which fixes the case where the glock number is
zero. This doesn't happen often, probably only for journal glocks,
although the fix is generic across glock types.

This has had a test as part of Abhi's recent testing for another 6.4
bug, and it is also easy to verify by inspection too.

Thanks are due to Nate Straz for spotting and fixing this bug. The patch
is queued for upstream, although it is not in the GFS2 git tree yet
since Linus has not yet released -rc1, but it will be included in due
course.

This fixes bug #884822

Signed-off-by: Jarod Wilson <jarod@redhat.com>