autofs4: use simple_empty() for empty directory check
Patchwork-id: 55948
O-Subject: [RHEL-6.4 PATCH 2/2] autofs4 - use simple_empty() for empty directory check
Bugzilla: 876795
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

Bugzilla bug: https://bugzilla.redhat.com/show_bug.cgi?id=876795

Upstream commit: 0259cb02c4004d3088b0999799f8f5c8801f6b97

edit: imk
This patch is not quite the same as the upstream commit because
the dcache_lock is still used by the rhel kernel. Consequently
some of the code simplifications aren't possible.
end edit: imk

For direct (and offset) mounts, if an automounted mount is manually
umounted the trigger mount dentry can appear non-empty causing it to
not trigger mounts. This can also happen if there is a file handle
leak in a user space automounting application.

It happens because, when the ioctl control file handle is opened
on the mount, a cursor dentry is created which causes list_empty()
to see the dentry as non-empty. Since there is a case where listing
the directory of these dentrys is needed, the use of dcache_dir_*()
functions for .open() and .release() is needed.

Consequently simple_empty() must be used instead of list_empty()
when checking for an empty directory.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Jarod Wilson <jarod@redhat.com>