jbd, jbd2: don't wake kjournald unnecessarily
Patchwork-id: 55967
O-Subject: [RHEL6 PATCH] jbd/jbd2: don't wake kjournald unnecessarily
Bugzilla: 886318
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

This is for
BZ 886318: AIM7 (shared workload) runs slower with ext3/ext4 than with earlier RHEL6 releases

It turns out that this is a regression caused by post-RHEL6.3
changes to __log_start_commit for bug 735768,

- [fs] jbd2: fix fsync() tid wraparound bug (Dave Wysochanski) [735768]
- [fs] jbd: fix fsync() tid wraparound bug (Dave Wysochanski) [735768]

The same problem exists upstream, so the backport caught it as well.

Barry has tested the following patch and found that it resolves
the perf problem.

It's been sent upstream & reviewed by Jan but not yet merged.

===

jbd/jbd2: don't wake kjournald unnecessarily

Don't send an extra wakeup to kjournald in the case where we
already have the proper target in j_commit_request, i.e. that
transaction has already been requested for commit.

commit deeeaf13 "jbd2: fix fsync() tid wraparound bug" changed
the logic leading to a wakeup, but it caused some extra wakeups
which were found to lead to a measurable performance regression.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Jan Kara <jack@suse.cz>