raid10: Do not call md_raid10_unplug_device while holding spinlock
Patchwork-id: 56003
O-Subject: [RHEL 6.4 PATCH] Do not call md_raid10_unplug_device while holding spinlock
Bugzilla: 886658
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

Bug 886658:
 https://bugzilla.redhat.com/show_bug.cgi?id=886658

Brew Build:
 http://brewweb.devel.redhat.com/brew/taskinfo?taskID=5210218

Testing:
 Create a RAID10 LV, kill one of the mirror pairs and attempt
a 'dd' write.  The system hangs and eventually is rebooted by
watchdog if this patch is not present.  I/O properly fails and
does not hang when the patch is in.  FS stress tests as well as
simple sequential I/O was confirmed to work as expected.

Upstream status:
 N/A - the plugging infrastructure is different upstream.

Patch Description:
 Don't call 'md_raid10_unplug_device' while holding a spinlock.
This is exactly the same problem seen for RAID1 in bz752528 -
commit 9617929.  Again, it is not necessary to hold the spinlock
in order to call the function and since it can block, it can
cause deadlocks if we do.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>