vfs: remove unneeded permission check from path_init
Patchwork-id: 56074
O-Subject: [RHEL6.5 PATCH] BZ#884863: vfs: remove unneeded permission check from path_init
Bugzilla: 884863
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

This patch eliminates an unneeded exec permission check from the path
walking code. This patch is a little different from the upstream code
due to differences in the pathwalking code. The effect should be the
same however.

Essentially, we want to skip the permissions check in path_init in
most cases since it'll be repeated soon afterward in __link_path_walk.
In the case of RHEL6, we must continue to check permissions here if
the pathname is empty since __link_path_walk will skip checking them
in that case.

Original patch description follows...

-----------------------[snip]-----------------------

When path_init is called with a valid dfd, that code checks permissions
on the open directory fd and returns an error if the check fails. This
permission check is redundant, however.

Both callers of path_init immediately call link_path_walk afterward. The
first thing that link_path_walk does for pathnames that do not consist
only of slashes is to check for exec permissions at the starting point of
the path walk.  And this check in path_init() is on the path taken only
when *name != '/' && *name != '\0'.

In most cases, these checks are very quick, but when the dfd is for a
file on a NFS mount with the actimeo=0, each permission check goes
out onto the wire. The result is 2 identical ACCESS calls.

Given that these codepaths are fairly "hot", I think it makes sense to
eliminate the permission check in path_init and simply assume that the
caller will eventually check the permissions before proceeding.

Reported-by: Dave Wysochanski <dwysocha@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Upstream commit: 582aa64a04a579d47d05e4a0ee85bf047978ef4d

Signed-off-by: Jarod Wilson <jarod@redhat.com>